+++ src/main/java/warmup/one/WarmupOne.java
@@ -28,4 +28,15 @@ public class WarmupOne {
     public boolean monkeyTrouble(boolean aSmile, boolean bSmile) {
         return aSmile && bSmile || !aSmile && !bSmile;
     }
+
+    /**
+     * Given two int values, return their sum. Unless the two values are the same, then return double their sum.
+     *
+     * @param a
+     * @param b
+     * @return int
+     */
+    public int sumDouble(int a, int b) {
+        return a == b ? (a+b) * 2 : a+b;
+    }
 }