+++ CHANGELOG.md
@@ -10,3 +10,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Warmup-1 sleepIn
 - Warmup-1 monkeyTrouble
+- Warmup-1 sumDouble